requirements.txt
# Core Dependencies
streamlit==1.51.0
openai==2.6.1
python-dotenv==1.2.1
pydantic==2.12.3

# Database
snowflake-connector-python==4.0.0
pandas==2.3.3

# Visualization
plotly==6.3.1
altair==5.5.0

# Security & ML
transformers==4.57.1
torch==2.9.0
huggingface-hub==0.36.0

# Streamlit Extensions
streamlit-extras==0.7.8
streamlit-camera-input-live==0.2.0

# Utilities
requests==2.32.5
scipy>=1.11.0
